requirements.txt
numpy #TODO ADD Fixed Versions
pandas
nba_api
request